relocate parsemetadatacolls to exchange
