+++ src/internal/connector/data_collections.go
@@ -2,7 +2,6 @@ package connector
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"strings"
 
@@ -33,7 +32,7 @@ import (
 func (gc *GraphConnector) DataCollections(
 	ctx context.Context,
 	sels selectors.Selector,
-	metadataCols []data.Collection,
+	metadata []data.Collection,
 ) ([]data.Collection, error) {
 	ctx, end := D.Span(ctx, "gc:dataCollections", D.Index("service", sels.Service.String()))
 	defer end()
@@ -43,14 +42,9 @@ func (gc *GraphConnector) DataCollections(
 		return nil, err
 	}
 
-	_, deltas, err := parseMetadataCollections(ctx, metadataCols)
-	if err != nil {
-		return nil, err
-	}
-
 	switch sels.Service {
 	case selectors.ServiceExchange:
-		return gc.ExchangeDataCollection(ctx, sels, deltas)
+		return gc.ExchangeDataCollection(ctx, sels, metadata)
 	case selectors.ServiceOneDrive:
 		return gc.OneDriveDataCollections(ctx, sels)
 	case selectors.ServiceSharePoint:
@@ -111,56 +105,6 @@ func verifyBackupInputs(sels selectors.Selector, userPNs, siteIDs []string) erro
 	}
 
 	return nil
-}
-
-// parseMetadataCollections produces two maps:
-// 1- paths: folderID->filePath, used to look up previous folder pathing
-// in case of a name change or relocation.
-// 2- deltas: folderID->deltaToken, used to look up previous delta token
-// retrievals.
-func parseMetadataCollections(
-	ctx context.Context,
-	colls []data.Collection,
-) (map[string]string, map[string]string, error) {
-	var (
-		paths  = map[string]string{}
-		deltas = map[string]string{}
-	)
-
-	for _, coll := range colls {
-		items := coll.Items()
-
-		for {
-			var breakLoop bool
-
-			select {
-			case <-ctx.Done():
-				return nil, nil, errors.Wrap(ctx.Err(), "parsing collection metadata")
-			case item, ok := <-items:
-				if !ok {
-					breakLoop = true
-					break
-				}
-
-				switch item.UUID() {
-				// case graph.PreviousPathFileName:
-				case graph.DeltaTokenFileName:
-					err := json.NewDecoder(item.ToReader()).Decode(&deltas)
-					if err != nil {
-						return nil, nil, errors.New("parsing delta token map")
-					}
-
-					breakLoop = true
-				}
-			}
-
-			if breakLoop {
-				break
-			}
-		}
-	}
-
-	return paths, deltas, nil
 }
 
 // ---------------------------------------------------------------------------
@@ -234,7 +178,7 @@ func (gc *GraphConnector) createExchangeCollections(
 func (gc *GraphConnector) ExchangeDataCollection(
 	ctx context.Context,
 	selector selectors.Selector,
-	deltas map[string]string,
+	metadata []data.Collection,
 ) ([]data.Collection, error) {
 	eb, err := selector.ToExchangeBackup()
 	if err != nil {
@@ -246,6 +190,11 @@ func (gc *GraphConnector) ExchangeDataCollection(
 		collections = []data.Collection{}
 		errs        error
 	)
+
+	_, deltas, err := exchange.ParseMetadataCollections(ctx, metadata)
+	if err != nil {
+		return nil, err
+	}
 
 	for _, scope := range scopes {
 		// Creates a map of collections based on scope
